requirements.txt
# Core dependencies - pinned for compatibility
numpy<2.0  # Required for PyTorch compatibility
torch>=2.2.0  # sentence-transformers requires >= 2.2
transformers>=4.36.0,<5.0.0
datasets>=2.16.0
sentence-transformers>=2.2.0

# Evaluation
bert-score>=0.3.13
nltk>=3.8.0
sacrebleu>=2.4.0

# Utilities
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Vector search
faiss-cpu>=1.7.4

# Optional: for 4-bit/8-bit quantization (reduces memory usage)
# bitsandbytes>=0.41.0

# Optional: for GPU support
# faiss-gpu>=1.7.4
